+++ docs/source/conf.py
@@ -26,7 +26,7 @@ author = 'Daniel Estévez'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '3.5.2'
+release = '3.6.0'
 
 
 # -- General configuration ---------------------------------------------------