+++ lib/ax100_decode_impl.cc
@@ -44,8 +44,7 @@ ax100_decode_impl::ax100_decode_impl(bool verbose)
 {
     message_port_register_out(pmt::mp("out"));
     message_port_register_in(pmt::mp("in"));
-    set_msg_handler(pmt::mp("in"),
-                    boost::bind(&ax100_decode_impl::msg_handler, this, _1));
+    set_msg_handler(pmt::mp("in"), [this](pmt::pmt_t msg) { this->msg_handler(msg); });
 }
 
 /*