+++ lib/decode_rs_impl.cc
@@ -15,40 +15,95 @@
 #include "decode_rs_impl.h"
 #include <gnuradio/io_signature.h>
 
-#include <cstdio>
+#include <boost/format.hpp>
+#include <algorithm>
+#include <exception>
 
 extern "C" {
 #include "libfec/fec.h"
 }
 
-#include "rs.h"
-
 namespace gr {
 namespace satellites {
 
-decode_rs::sptr decode_rs::make(bool verbose, int basis)
+decode_rs::sptr decode_rs::make(int dual_basis, int interleave)
 {
-    return gnuradio::get_initial_sptr(new decode_rs_impl(verbose, basis));
+    return gnuradio::get_initial_sptr(new decode_rs_impl(dual_basis, interleave));
+}
+
+decode_rs::sptr
+decode_rs::make(int symsize, int gfpoly, int fcr, int prim, int nroots, int interleave)
+{
+    return gnuradio::get_initial_sptr(
+        new decode_rs_impl(symsize, gfpoly, fcr, prim, nroots, interleave));
 }
 
 /*
  * The private constructor
  */
-decode_rs_impl::decode_rs_impl(bool verbose, int basis)
+decode_rs_impl::decode_rs_impl(bool dual_basis, int interleave)
     : gr::block(
           "decode_rs", gr::io_signature::make(0, 0, 0), gr::io_signature::make(0, 0, 0)),
-      d_verbose(verbose),
-      d_basis(basis)
+      d_interleave(interleave)
+{
+    if (dual_basis) {
+        d_decode_rs = [](uint8_t* data) { return decode_rs_ccsds(data, NULL, 0, 0); };
+    } else {
+        d_decode_rs = [](uint8_t* data) { return decode_rs_8(data, NULL, 0, 0); };
+    }
+    d_rs_codeword.resize(d_ccsds_nn);
+    d_nroots = d_ccsds_nroots;
+
+    check_interleave();
+    set_message_ports();
+}
+
+/*
+ * The private constructor
+ */
+decode_rs_impl::decode_rs_impl(
+    int symsize, int gfpoly, int fcr, int prim, int nroots, int interleave)
+    : gr::block(
+          "decode_rs", gr::io_signature::make(0, 0, 0), gr::io_signature::make(0, 0, 0)),
+      d_interleave(interleave)
+{
+    d_rs_p = init_rs_char(symsize, gfpoly, fcr, prim, nroots, 0);
+    if (!d_rs_p) {
+        throw std::runtime_error("Unable to initialize Reed-Solomon definition");
+    }
+    d_decode_rs = [this](uint8_t* data) { return decode_rs_char(d_rs_p, data, 0, 0); };
+
+    d_rs_codeword.resize((1U << symsize) - 1);
+    d_nroots = nroots;
+
+    check_interleave();
+    set_message_ports();
+}
+
+void decode_rs_impl::check_interleave()
+{
+    if (d_interleave <= 0) {
+        throw std::runtime_error(
+            boost::str(boost::format("Invalid interleave value = %d") % d_interleave));
+    }
+}
+
+void decode_rs_impl::set_message_ports()
 {
     message_port_register_out(pmt::mp("out"));
     message_port_register_in(pmt::mp("in"));
-    set_msg_handler(pmt::mp("in"), boost::bind(&decode_rs_impl::msg_handler, this, _1));
+    set_msg_handler(pmt::mp("in"), [this](pmt::pmt_t msg) { this->msg_handler(msg); });
 }
 
 /*
  * Our virtual destructor.
  */
-decode_rs_impl::~decode_rs_impl() {}
+decode_rs_impl::~decode_rs_impl()
+{
+    if (d_rs_p) {
+        free_rs_char(d_rs_p);
+    }
+}
 
 void decode_rs_impl::forecast(int noutput_items, gr_vector_int& ninput_items_required) {}
 
@@ -62,40 +117,61 @@ int decode_rs_impl::general_work(int noutput_items,
 
 void decode_rs_impl::msg_handler(pmt::pmt_t pmt_msg)
 {
-    size_t length(0);
-    auto msg = pmt::u8vector_elements(pmt::cdr(pmt_msg), length);
-    int rs_res;
+    auto msg = pmt::u8vector_elements(pmt::cdr(pmt_msg));
+    int errors = 0;
 
-    if (length <= PARITY_BYTES || length > d_data.size()) {
-        if (d_verbose) {
-            std::printf("Reed-Solomon decoder: invalid frame length %ld\n", (long)length);
-        }
+    if (msg.size() % d_interleave != 0) {
+        GR_LOG_WARN(d_logger,
+                    boost::format("Reed-Solomon message size not divisible by interleave "
+                                  "depth. size = %d, interleave = %d") %
+                        msg.size() % d_interleave);
         return;
     }
 
-    memcpy(d_data.data(), msg, length);
-
-    if (d_basis == BASIS_CONVENTIONAL) {
-        rs_res = decode_rs_8(d_data.data(), NULL, 0, MAX_FRAME_LEN - length);
-    } else {
-        rs_res = decode_rs_ccsds(d_data.data(), NULL, 0, MAX_FRAME_LEN - length);
+    int rs_nn = msg.size() / d_interleave;
+    if (rs_nn <= d_nroots || (unsigned)rs_nn > d_rs_codeword.size()) {
+        GR_LOG_ERROR(
+            d_logger,
+            boost::format("Wrong Reed-Solomon message size. size = %d, interleave "
+                          "= %d, RS code (%d, %d)") %
+                msg.size() % d_interleave % d_rs_codeword.size() %
+                (d_rs_codeword.size() - d_nroots));
+        return;
     }
 
-    // Send via GNUradio message if RS ok
-    if (rs_res >= 0) {
-        length -= PARITY_BYTES;
+    d_output_frame.resize(msg.size() - d_interleave * d_nroots);
+    const auto pad = d_rs_codeword.size() - rs_nn;
 
-        if (d_verbose) {
-            std::printf("Reed-Solomon decode OK. Bytes corrected: %d.\n", rs_res);
+    for (int j = 0; j < d_interleave; ++j) {
+        std::fill(d_rs_codeword.begin(), d_rs_codeword.begin() + pad, 0);
+        for (int k = 0; k < rs_nn; ++k) {
+            d_rs_codeword[pad + k] = msg[j + k * d_interleave];
         }
 
-        // Send by GNUradio message
-        message_port_pub(
-            pmt::mp("out"),
-            pmt::cons(pmt::PMT_NIL, pmt::init_u8vector(length, d_data.data())));
-    } else if (d_verbose) {
-        std::printf("Reed-Solomon decode failed.\n");
+        auto rs_res = d_decode_rs(d_rs_codeword.data());
+        if (rs_res < 0) {
+            GR_LOG_DEBUG(d_logger,
+                         boost::format("Reed-Solomon decode fail (interleaver path %d)") %
+                             j);
+            return;
+        }
+        GR_LOG_DEBUG(d_logger,
+                     boost::format(
+                         "Reed-Solomon decode corrected %d bytes (interleaver path %d)") %
+                         rs_res % j);
+        errors += rs_res;
+
+        for (int k = 0; k < rs_nn - d_nroots; ++k) {
+            d_output_frame[j + k * d_interleave] = d_rs_codeword[pad + k];
+        }
     }
+
+    auto meta =
+        pmt::dict_add(pmt::car(pmt_msg), pmt::mp("rs_errors"), pmt::from_long(errors));
+
+    message_port_pub(
+        pmt::mp("out"),
+        pmt::cons(meta, pmt::init_u8vector(d_output_frame.size(), d_output_frame)));
 }
 
 } /* namespace satellites */