+++ lib/matrix_deinterleaver_soft_impl.cc
@@ -50,8 +50,7 @@ matrix_deinterleaver_soft_impl::matrix_deinterleaver_soft_impl(int rows,
 
     message_port_register_out(pmt::mp("out"));
     message_port_register_in(pmt::mp("in"));
-    set_msg_handler(pmt::mp("in"),
-                    boost::bind(&matrix_deinterleaver_soft_impl::msg_handler, this, _1));
+    set_msg_handler(pmt::mp("in"), [this](pmt::pmt_t msg) { this->msg_handler(msg); });
 }
 
 /*