+++ lib/u482c_decode_impl.cc
@@ -69,8 +69,7 @@ u482c_decode_impl::u482c_decode_impl(bool verbose, int viterbi, int scrambler, i
 
     message_port_register_out(pmt::mp("out"));
     message_port_register_in(pmt::mp("in"));
-    set_msg_handler(pmt::mp("in"),
-                    boost::bind(&u482c_decode_impl::msg_handler, this, _1));
+    set_msg_handler(pmt::mp("in"), [this](pmt::pmt_t msg) { this->msg_handler(msg); });
 }
 
 /*