+++ python/components/deframers/aistechsat_2_deframer.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-# Copyright 2019 Daniel Estevez <daniel@destevez.net>
+# Copyright 2020 Daniel Estevez <daniel@destevez.net>
 #
 # This file is part of gr-satellites
 #
@@ -9,16 +9,16 @@
 #
 
 from gnuradio import gr, digital
-from ... import decode_rs_interleaved
+from ... import decode_rs, pdu_head_tail
 from ...hier.sync_to_pdu import sync_to_pdu
 from ...hier.ccsds_descrambler import ccsds_descrambler
 from ...utils.options_block import options_block
 
-_syncword = '00011010110011111111110000011101'
+from .ccsds_rs_deframer import _syncword
 
-class astrocast_9k6_deframer(gr.hier_block2, options_block):
+class aistechsat_2_deframer(gr.hier_block2, options_block):
     """
-    Hierarchical block to deframe the Astrocast custom 9k6 framing
+    Hierarchical block to deframe AISTECHSAT-2 custom frames
 
     The input is a float stream of soft symbols. The output are PDUs
     with frames.
@@ -28,34 +28,35 @@ class astrocast_9k6_deframer(gr.hier_block2, options_block):
         options: Options from argparse
     """
     def __init__(self, syncword_threshold = None, options = None):
-        gr.hier_block2.__init__(self, "astrocast_9k6_deframer",
+        gr.hier_block2.__init__(self, "aistechast_2_deframer",
             gr.io_signature(1, 1, gr.sizeof_float),
             gr.io_signature(0, 0, 0))
         options_block.__init__(self, options)
-        
+
         self.message_port_register_hier_out('out')
 
         if syncword_threshold is None:
             syncword_threshold = self.options.syncword_threshold
-
+        
         self.slicer = digital.binary_slicer_fb()
-        self.deframer = sync_to_pdu(packlen = 8 * 255 * 5,\
-                                    sync = _syncword,\
+        self.deframer = sync_to_pdu(packlen = (93 + 32) * 8,
+                                    sync = _syncword,
                                     threshold = syncword_threshold)
         self.scrambler = ccsds_descrambler()
-        self.rs = decode_rs_interleaved(self.options.verbose_rs, 1, 5)
+        self.tail = pdu_head_tail(3, 10)
+        self.fec = decode_rs(False, 1)
 
         self.connect(self, self.slicer, self.deframer)
         self.msg_connect((self.deframer, 'out'), (self.scrambler, 'in'))
-        self.msg_connect((self.scrambler, 'out'), (self.rs, 'in'))
-        self.msg_connect((self.rs, 'out'), (self, 'out'))
+        self.msg_connect((self.scrambler, 'out'), (self.tail, 'in'))
+        self.msg_connect((self.tail, 'out'), (self.fec, 'in'))
+        self.msg_connect((self.fec, 'out'), (self, 'out'))
 
     _default_sync_threshold = 4
         
     @classmethod
     def add_options(cls, parser):
         """
-        Adds Astrocast 9k6 deframer specific options to the argparse parser
+        Adds AISTECHSAT-2 deframer specific options to the argparse parser
         """
         parser.add_argument('--syncword_threshold', type = int, default = cls._default_sync_threshold, help = 'Syncword bit errors [default=%(default)r]')
-        parser.add_argument('--verbose_rs', action = 'store_true', help = 'Verbose RS decoder')