+++ python/sx12xx_packet_crop.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-# Copyright 2019 Daniel Estevez <daniel@destevez.net>
+# Copyright 2020 jgromes <gromes.jan@gmail.com>
 #
 # This file is part of gr-satellites
 #
@@ -12,19 +12,18 @@ import numpy
 from gnuradio import gr
 import pmt
 
-class header_remover(gr.basic_block):
+class sx12xx_packet_crop(gr.basic_block):
     """
-    Removes some bytes from the beginning of a PDU
-
-    Args:
-        length: Length to remove (int)
+    Crop SX12xx FSK packet, based on internal packet length field (first byte after sync word).
     """
-    def __init__(self, length):
+    def __init__(self, crc_len):
         gr.basic_block.__init__(self,
-            name="header_remover",
+            name="sx12xx_packet_crop",
             in_sig=[],
             out_sig=[])
-        self.length = length
+
+        self.crc_len = crc_len
+        
         self.message_port_register_in(pmt.intern('in'))
         self.set_msg_handler(pmt.intern('in'), self.handle_msg)
         self.message_port_register_out(pmt.intern('out'))
@@ -36,10 +35,6 @@ class header_remover(gr.basic_block):
             return
         packet = bytes(pmt.u8vector_elements(msg))
 
-        if len(packet) <= self.length:
-            return
-
-        data = packet[self.length:]
-        self.message_port_pub(pmt.intern('out'),
-                                  pmt.cons(pmt.car(msg_pmt),
-                                           pmt.init_u8vector(len(data), data)))
+        packet_length = packet[0] + 1 + self.crc_len
+        
+        self.message_port_pub(pmt.intern('out'),  pmt.cons(pmt.car(msg_pmt), pmt.init_u8vector(packet_length, packet)))