+++ CHANGELOG.md
@@ -4,6 +4,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [3.6.0] - 2020-12-04
+
+### Added
+- PDU add metadata block
+- PDU Head/Tail block
+- Lucky-7 image receiver
+- Support for custom SIDS servers
+- Support for AISTECHSAT-2 custom protocol
+- Support for BOBCAT-1
+- Support for BY03
+- Support for FossaSat-1B and FossaSat-2
+- Support for NEUTRON-1
+- Support for SPOC
+- Support for TTU-100
+- Support for VZLUSAT-2
+
+### Changed
+- Replaced boost::bind() by C++ lambdas
+- Refactored Reed-Solomon decoder C++ blocks
+- Refactored CCSDS deframer blocks to allow more generality
+- Make --clk_limit parameter relative to samples per symbol
+- Updated AISTECHSAT-2 transmit frequency
+- Definitive NORAD IDs for NETSAT 1-4
+
+### Removed
+- Deprecated Astrocast 9k6 deframer in favour of the new CCSDS Reed-Solomon deframer
+- Deprecated CC11xx remove length, Header remover and Strip AX.25 header in favour of PDU Head/Tail
+
 ## [3.5.2] - 2020-10-21
 
 ### Changed
@@ -288,7 +318,8 @@ Large refactor release bringing new functionality and improvements. This is an o
 ## [1.0.0] - 2018-08-02
 First gr-satellites release using semantic versioning
 
-[Unreleased]: https://github.com/daniestevez/gr-satellites/compare/v3.5.2...master
+[Unreleased]: https://github.com/daniestevez/gr-satellites/compare/v3.6.0...master
+[3.6.0]: https://github.com/daniestevez/gr-satellites/compare/v3.5.2...v3.6.0
 [3.5.2]: https://github.com/daniestevez/gr-satellites/compare/v3.5.1...v3.5.2
 [3.5.1]: https://github.com/daniestevez/gr-satellites/compare/v3.5.0...v3.5.1
 [3.5.0]: https://github.com/daniestevez/gr-satellites/compare/v3.4.0...v3.5.0